backend/requirements.txt
alembic==1.7.3
bcrypt==3.2.0
email-validator==1.1.3
fastapi==0.68.1
fastapi-pagination==0.9.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.1
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pyyaml==5.4.1
requests==2.26.0
semver==2.13.0
SQLAlchemy==1.4.23
types-python-dateutil==2.8.2